+++ demisto_sdk/commands/validate/validate_manager.py
@@ -56,7 +56,6 @@ class ValidateManager:
         self.no_configuration_prints = silence_init_prints
         self.skip_conf_json = skip_conf_json
         self.is_backward_check = is_backward_check
-        self.id_set_validations = validate_id_set
         self.is_circle = only_committed_files
         self.validate_all = validate_all
         self.use_git = use_git
@@ -72,9 +71,14 @@ class ValidateManager:
         # Class constants
         self.handle_error = BaseValidator(print_as_warnings=print_ignored_errors).handle_error
         self.file_path = file_path
-        if not id_set_path:
-            id_set_path = 'Tests/id_set.json'
-        self.id_set_path = id_set_path
+        self.id_set_path = id_set_path or IDSetValidator.ID_SET_PATH
+        # create the id_set only once per run.
+        self.id_set_validator = IDSetValidator(is_circle=self.is_circle,
+                                               configuration=Configuration(),
+                                               ignored_errors=None,
+                                               print_as_warnings=self.print_ignored_errors,
+                                               id_set_path=self.id_set_path) \
+            if validate_id_set else None
 
         try:
             self.git_util = GitUtil(repo=Content.git())
@@ -87,6 +91,7 @@ class ValidateManager:
                 self.branch_name = ''
                 pass
 
+        self.branch_name = ''
         self.changes_in_schema = False
         self.check_only_schema = False
         self.always_valid = False
@@ -317,12 +322,9 @@ class ValidateManager:
                 return valid_schema
 
         # id_set validation
-        if self.id_set_validations:
-            id_set_validator = IDSetValidator(is_circle=self.is_circle, configuration=Configuration(),
-                                              ignored_errors=pack_error_ignore_list,
-                                              print_as_warnings=self.print_ignored_errors)
-            if not id_set_validator.is_file_valid_in_set(file_path, file_type):
-                return False
+        if self.id_set_validator and not \
+                self.id_set_validator.is_file_valid_in_set(file_path, file_type):
+            return False
 
         elif file_type == FileType.RELEASE_NOTES:
             if not self.skip_pack_rn_validation: